Split vote handlers into separate files
